Restart queue container on failure
This should help to mitigate the race condition between the queue
container and the database container, as the queue application
will panic if it cannot establish a database connection, and thus
will keep restarting until it establishes one. This isn't super
clean, but I think it's the easiest immediate mitigation here.